--- app/includes/kwicresult.pug
@@ -1,63 +1,0 @@
-.result_controls
-  warning(ng-if="aborted && !loading")
-    | {{'search_aborted' | loc:lang}}
-  .controls_n
-    span {{'num_results' | loc:lang}}: 
-    span.num-result(ng-show="!loading && hits_display" ng-bind-html="hits_display | trust")
-  .hits_picture(ng-if='hitsPictureData.length > 1')
-    table.hits_picture_table
-      tbody
-        tr
-          td.hits_picture_corp(ng-repeat='corpus in hitsPictureData'
-              ng-style='{width : corpus.relative + "%"}'
-              ng-class='{odd : $index % 2 != 0, \
-                         even : $index % 2 == 0}'
-              ng-click='hitspictureClick(corpus.page, corpus)'
-              uib-tooltip='{{corpus.rtitle | locObj:lang}}: {{corpus.abs}}'
-              uib-tooltip-placement='{{$last? "left":"top"}}'
-              append-to-body="false")
-kwic-pager(ng-if="hits" total-hits="hits", current-page="page", page-change="pageChange(page)", hits-per-page="pagerHitsPerPage")
-
-span.reading_btn.show_reading.link(ng-hide="loading", ng-click="toggleReading()") {{'show_reading' | loc:lang}}
-span.reading_btn.hide_reading.link(ng-hide="loading", ng-click="toggleReading()") {{'show_kwic' | loc:lang}}
-.table_scrollarea(ng-class="{paging : paging}")
-  table.results_table.kwic(ng-if="kwic.length > 0" cellspacing='0')
-    tr.sentence(ng-repeat='sentence in kwic'
-      ng-class='{corpus_info : sentence.newCorpus, \
-        not_corpus_info : !sentence.newCorpus, \
-        linked_sentence : sentence.isLinked, \
-        even : $even, \
-        odd : $odd}')
-      td.empty_td
-      td.corpus_title.text-gray-600.uppercase.text-sm(colspan='3')
-        div
-          | {{sentence.newCorpus | locObj:lang}}
-          span.corpus_title_warn(rel='localize[no_context_support]', ng-if='::sentence.noContext')
-      td.empty_td
-      td.lnk(colspan='3', ng-if='::sentence.isLinked')
-        span(kwic-word='kwic-word', ng-repeat='wd in sentence.tokens')
-      td.left(ng-if='::!sentence.newCorpus')
-        span(kwic-word='kwic-word', ng-repeat='wd in selectLeft(sentence)')
-      td.match(ng-if='::!sentence.newCorpus')
-        span(kwic-word='kwic-word', ng-repeat='wd in selectMatch(sentence)')
-      td.right(ng-if='::!sentence.newCorpus')
-        span(kwic-word='kwic-word', ng-repeat='wd in selectRight(sentence)')
-  .results_table.reading(ng-if="contextKwic.length > 0")
-    p.sentence(ng-repeat='sentence in contextKwic'
-      ng-class='{corpus_info : sentence.newCorpus, \
-        not_corpus_info : !sentence.newCorpus, \
-        linked_sentence : sentence.isLinked, \
-        even : $even, \
-        odd : $odd}')
-      span.corpus_title(colspan='0')
-        | {{sentence.newCorpus | locObj:lang}}
-        span.corpus_title_warn(rel='localize[no_context_support]', ng-if='::sentence.noContext')
-      span(ng-repeat='wd in sentence.tokens', kwic-word='kwic-word')
-kwic-pager(ng-if="hits" total-hits="hits", current-page="page", page-change="pageChange(page)", hits-per-page="pagerHitsPerPage")
-#download-links-container
-  select#download-links(ng-if="_settings['enable_backend_kwic_download']")
-  select#frontendDownloadLinks(ng-if="_settings['enable_frontend_kwic_download']",
-                        ng-change="download.init(download.selected, hits_display)", 
-                        ng-model="download.selected",
-                        ng-options="item.value as item.label | loc:lang disable when item.disabled for item in download.options")
-  a.kwicDownloadLink(ng-if="_settings['enable_frontend_kwic_download']", href="{{download.blobName}}", download="{{download.fileName}}", target="_self", style="display: none;")
