+++ app/index.js
@@ -97,7 +97,6 @@ require("./scripts/cqp_parser/CQPParser.js")
 require("./scripts/cqp_parser/cqp.js")
 require("./scripts/util.js")
 require("./scripts/pie-widget.js")
-require("./scripts/results/results.js")
 require("./scripts/model.js")
 require("./scripts/widgets.js")
 require("./scripts/main.js")
@@ -105,7 +104,15 @@ require("./scripts/app.js")
 require("./scripts/search_controllers.js")
 
 require("./scripts/kwic_download.js")
-require("./scripts/result_controllers.js")
+
+require("./scripts/results/controllers/result_controllers.js")
+require("./scripts/results/controllers/comparison_controller.js")
+require("./scripts/results/controllers/kwic_controller.js")
+require("./scripts/results/controllers/example_controller.js")
+require("./scripts/results/controllers/statistics_controller.js")
+require("./scripts/results/controllers/trend_diagram_controller.js")
+require("./scripts/results/controllers/word_picture_controller.js")
+
 require("./scripts/map_controllers.js")
 require("./scripts/text_reader_controller.js")
 require("./scripts/video_controllers.js")
@@ -113,6 +120,7 @@ require("./scripts/services.js")
 require("./scripts/extended.js")
 require("./scripts/struct_services.js")
 require("./scripts/directives.js")
+require("./scripts/directives2.js")
 require("./scripts/directives/scroll.js")
 require("./scripts/filter_directives.js")
 require("./scripts/newsdesk.js")
