+++ app/scripts/components/simple_search.js
@@ -251,40 +251,38 @@ export const simpleSearchComponent = {
                 const search = searches.activeSearch
                 ctrl.relatedObj = null
                 const cqp = ctrl.getCQP()
-                searches.kwicSearch(cqp, search && search.pageOnly)
+                searches.kwicSearch(cqp)
 
-                if (!(search && search.pageOnly)) {
-                    if (search.type === "lemgram") {
-                        let sense = false
-                        let saldo = false
-                        for (let corpus of settings.corpusListing.selected) {
-                            if ("sense" in corpus.attributes) {
-                                sense = true
-                            }
-                            if ("saldo" in corpus.attributes) {
-                                saldo = true
-                            }
+                if (search.type === "lemgram") {
+                    let sense = false
+                    let saldo = false
+                    for (let corpus of settings.corpusListing.selected) {
+                        if ("sense" in corpus.attributes) {
+                            sense = true
                         }
-
-                        if (sense || saldo) {
-                            backend.relatedWordSearch(unregescape(search.val)).then(function (data) {
-                                ctrl.relatedObj = data
-                                if (data.length > 2 && data[0].label == "Excreting") {
-                                    let [first, second, ...rest] = data
-                                    ctrl.relatedObj.data = [second, first, ...rest]
-                                }
-                                ctrl.relatedObj.attribute = sense ? "sense" : "saldo"
-                            })
+                        if ("saldo" in corpus.attributes) {
+                            saldo = true
                         }
                     }
 
-                    if (ctrl.wordPic && (search.type === "lemgram" || !search.val.includes(" "))) {
-                        const value = search.type === "lemgram" ? unregescape(search.val) : search.val
-                        return searches.lemgramSearch(value, search.type)
-                    } else {
-                        if ("wordPictureResults" in window) {
-                            wordPictureResults.resetView()
-                        }
+                    if (sense || saldo) {
+                        backend.relatedWordSearch(unregescape(search.val)).then(function (data) {
+                            ctrl.relatedObj = data
+                            if (data.length > 2 && data[0].label == "Excreting") {
+                                let [first, second, ...rest] = data
+                                ctrl.relatedObj.data = [second, first, ...rest]
+                            }
+                            ctrl.relatedObj.attribute = sense ? "sense" : "saldo"
+                        })
+                    }
+                }
+
+                if (ctrl.wordPic && (search.type === "lemgram" || !search.val.includes(" "))) {
+                    const value = search.type === "lemgram" ? unregescape(search.val) : search.val
+                    return searches.lemgramSearch(value, search.type)
+                } else {
+                    if ("wordPictureResults" in window) {
+                        wordPictureResults.resetView()
                     }
                 }
             }
